improvement: @jonator feedback
