+++ packages/web/hooks/one-click-trading/use-one-click-trading-params.ts
@@ -43,18 +43,6 @@ function getParametersFromOneClickTradingInfo({
  * can be either the custom parameters passed at initialization or the fetched default parameters.
  *
  * This hook is primarily intended for the one click trading settings modal.
- *
- * @param {Object} params - The parameters object for initializing the hook.
- * @param {OneClickTradingInfo} params.oneClickTradingInfo - Optional. Custom initial parameters for one-click trading.
- * @param {boolean} params.defaultIsOneClickEnabled - Optional. Specifies if one-click trading is enabled by default.
- *                                                     Defaults to false if not provided.
- * @returns {Object} An object containing:
- *                    - `transaction1CTParams`: The current parameters for one-click trading transactions.
- *                    - `setTransaction1CTParams`: Function to manually set the transaction parameters.
- *                    - `spendLimitTokenDecimals`: The decimal precision for the spend limit token.
- *                    - `isLoading`: Boolean indicating if the default parameters are being fetched.
- *                    - `isError`: Boolean indicating if there was an error fetching the default parameters.
- *                    - `reset`: Function to reset the parameters to their initial values.
  */
 export const useOneClickTradingParams = ({
   oneClickTradingInfo,
